Account for the new I16 type in the text package.
--HG--
extra : convert_revision : e9397a616bb696fc937c440d266873b37d903d7a

+++ Data/Text/ICU/Text.hs
@@ -54,8 +54,8 @@ toCaseFold excludeI s = unsafePerformIO .
                                   (fromIntegral slen) opts
           if n > len
             then go n
-            else fromPtr dptr n
-    go slen
+            else fromPtr dptr (fromIntegral n)
+    go (fromIntegral slen)
 
 type CaseMapper = Ptr UChar -> Int32 -> Ptr UChar -> Int32 -> CString
                 -> Ptr UErrorCode -> IO Int32
@@ -70,8 +70,8 @@ caseMap mapFn loc s = unsafePerformIO .
                               (fromIntegral slen) locale
             if n > len
               then go n
-              else fromPtr dptr n
-      go slen
+              else fromPtr dptr (fromIntegral n)
+      go (fromIntegral slen)
 
 -- | Lowercase the characters in a string.
 --
